users: Improve db transaction structure in user (de)activation process.
These procedures should be done atomically overall, with the exception
of the code that sends events to avoid block if there's a delay
communicating with Tornado.
We add the savepoint=False on the two underlying functions that already
execute inside an atomic context - to avoid the overhead of creating
savepoints where they aren't needed.